Refactored Kunena system plugin into J6 compatible plugin
