+++ src/plugins/plg_system_kunena/script.php
@@ -12,11 +12,9 @@
 
 defined('_JEXEC') or die();
 
-use Joomla\CMS\Application\CMSApplication;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Installer\Adapter\ComponentAdapter;
 use Joomla\CMS\Installer\InstallerScript;
-use Joomla\Database\DatabaseDriver;
 use Joomla\Database\Exception\ExecutionFailureException;
 
 /**
@@ -40,7 +38,7 @@ class plgSystemKunenaInstallerScript extends InstallerScript
      * @var    string
      * @since  5.4.0
      */
-    protected $minimumPhp = '7.2.5';
+    protected $minimumPhp = '8.1';
 
     /**
      * Minimum Joomla! version required to install the extension
@@ -48,7 +46,7 @@ class plgSystemKunenaInstallerScript extends InstallerScript
      * @var    string
      * @since  6.0.0
      */
-    protected $minimumJoomla = '4.0.0-dev';
+    protected $minimumJoomla = '5.3.2';
 
     /**
      * List of required PHP extensions.
@@ -59,21 +57,14 @@ class plgSystemKunenaInstallerScript extends InstallerScript
     protected $extensions = ['dom', 'gd', 'json', 'pcre', 'SimpleXML'];
 
     /**
-     * @var  string  During an update, it will be populated with the old release version
-     *
-     * @since   Kunena 6.0
-     */
-    private $oldRelease;
-
-    /**
      * method to run after an install/update/uninstall method
      *
      * @param   string            $type    'install', 'update' or 'discover_install'
      * @param   ComponentAdapter  $parent  Installer object
      *
-     * @return void
+     * @return  void
      *
-     * @throws Exception
+     * @throws  Exception
      * @since   Kunena 6.0
      */
     public function postflight($type, $parent)
@@ -88,18 +79,19 @@ class plgSystemKunenaInstallerScript extends InstallerScript
      * @param   InstallerAdapter  $parent  The class calling this method
      *
      * @return  boolean  True on success
-     *
      * @since   Kunena 6.5
      */
-    public function preflight($type, $parent)
+    public function preflight($type, $parent): bool
     {
         if (!parent::preflight($type, $parent)) {
             return false;
         }
 
         // Delete kunena.php
-        $this->deleteFiles[JPATH_SITE . '/plugins/system/kunena/kunena.php'];
+        $this->deleteFiles[] = JPATH_SITE . '/plugins/system/kunena/kunena.php';
         $this->removeFiles();
+
+        return true;
     }
 
 
@@ -107,9 +99,9 @@ class plgSystemKunenaInstallerScript extends InstallerScript
      * @param $pluginName
      *
      * @return boolean|false
-     * @since version
+     * @since  version
      */
-    public function enablePlugin($pluginName)
+    public function enablePlugin(string $pluginName): bool
     {
         // Create a new db object.
         $db    = Factory::getContainer()->get('DatabaseDriver');
@@ -129,5 +121,7 @@ class plgSystemKunenaInstallerScript extends InstallerScript
         } catch (ExecutionFailureException $e) {
             return false;
         }
+
+        return true;
     }
 }
